Added support for hex/octal/binary literals.
Made all number literals be stored in binary.  The justifications for storing
them as text for JSON documents don't really apply for scripts.

Fixed a bug: functions that used a product as their argument were treating it
as "true".  This was a bug in the effort to make "count(*)" count all items.

Modified SELECT to assume DISTINCT if all columns use an aggregate function.
This is intended to make "SELECT count(*) FROM table" output a single count
instead of one for each row.