Extended the - operator to concatenate strings with a single space.
Extended the IN operator to treat a single-column table like an array of
values.  This is important for nested SELECT clauses.

Removed the last vestiges of allowing IN to use a parenthesized list of
literals (SQL-style) instead of a bracketted list of literals (JSON-style).
The abuse of parentheses this way messed up using parentheses around
more complex expressions that return lists.