Made it recognize -? as a synonym for --help.
Tightened the rules for when it assumes "-" as an input file if no files
are specified.  Now it will only do this if stdin is not a tty.  This
change was necessary because without the "not a tty" rule, a simple
command like "jsoncalc -c2+2" would hang, waiting for the user to type in
JSON data that would then be ignored.