This is a big change to the way positions of errors are tracked.  Previously the code in cmd.c used the filename and line number to track the position of each command, and used that in error messages but errors from the json_calc() expression evaluator don't have filename and line number.  Consequently, if an error was detected in a user-defined function then json_calc() had no way to indicate where; instead, the error was reported where the function was called, not the line within the function where the error was detected.
This change fixes that by keeping all script files mapped into memory and
using a (char *) to indicate where errors are detected.  The code in cmd.c
was extensively modified to use that.  Code in calc.c was modified too, but
it mostly returns NULL as the position which causes cmd.c to use the command's
position still... except for user-defined functions, which return position
where the error was detected.  That solves that problem!

Also removed the error "code" numbers.  Instead, each error message can be
optionally prefixed with by a "key:" to indicate which error message it is;
these keys can be used to look up a translated version of the error message.

Along the way, some "char *" variables became "const char *".  I'm happy that
I was able to tighten that up.