Changed all (I hope) output to use new json_user_printf() and json_user_ch() functions, instead of standard functions like printf() and putch().  This was done so a user interface can divert them -- necessary for a curses-based interface, and probably a GUI interface too.
Changed the way text colors/attributes are stored in the config.  They are
now part of a table named "config.styles".  This is related the the output
change, above.

Added a json_config_style() function for looking up (and maybe adding)
a text style.  Modified the "log" plugin to use it.