Implemented the parser in the -lxml plugin.  Renamed some of the source files for that plugin too.
Changed the json_parser_hook() function to take an extra parameter for an
"updater" function pointer.  That function writes the data back out.  The
idea (not fully implemented yet) is that if you load an XML file and then
try to update/write it, it should be written in XML too.  I modified the
-lxml and -lcsv parsers to use this.

I added some assert() statements in strategic places, especially in the
json_append() function.